Login, Logout, Slideshow, and others

+++ handlers/delete_albums.py
@@ -1,3 +1,5 @@
+from google.appengine.api import users
+
 __author__ = 'Binnie'
 
 import webapp2
@@ -9,12 +11,14 @@ from google.appengine.ext.webapp import template
 
 class DeleteAlbumsHandler(webapp2.RequestHandler):
     def get(self):
-        photos = Image.query().fetch()
-        albums = Album.query().fetch()
-        for photo in photos:
-            photo_key = ndb.Key(urlsafe=photo.ukey())
-            photo_key.delete()
-        for album in albums:
-            album_key = ndb.Key(urlsafe=album.ukey())
-            album_key.delete()
+        if users.is_current_user_admin():
+            photos = Image.query().fetch()
+            albums = Album.query().fetch()
+            for photo in photos:
+                photo_key = ndb.Key(urlsafe=photo.ukey())
+                photo_key.delete()
+            for album in albums:
+                album_key = ndb.Key(urlsafe=album.ukey())
+                album_key.delete()
+
         self.redirect('/')
